pyproject.toml
[tool.poetry]
name = "cross-stitch-tasks"
version = "0.1.0"
description = ""
authors = ["Alexander Gnusarev <alexander.gnussarev@gmail.com>"]
readme = "README.md"
packages = [{include = "cross_stitch_tasks", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
alembic = "^1.14.0"
psycopg2 = "^2.9.10"
flask-wtf = "^1.2.2"
pandas = "^2.2.3"
flask-migrate = "^4.0.7"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
pylint = "^3.3.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
